# feat(instance) Add a new `wasm_instance.exported_functions` field (improve function invocation time by 77%)