--- a/extension/wasm.hh
+++ b/extension/wasm.hh
@@ -26,6 +26,8 @@
 #include "Zend/zend_interfaces.h"
 #include "php_wasm.h"
 #include "wasmer.hh"
+#include <unordered_map>
+#include <string>
 
 #if defined(PHP_WIN32)
 #  include "win32/php_stdint.h"
@@ -158,7 +160,16 @@ typedef struct {
 
     // The internal opaque exports pointer.
     wasmer_exports_t *exports;
+
+    // A map from exported function names to exported function pointers.
+    std::unordered_map<std::string, const wasmer_export_func_t *> *exported_functions;
 } wasm_instance;
+
+/**
+ * Given an already instantiated `wasm_instance`, this function will
+ * fill the other fields (`exports` and `exported_functions`).
+ */
+void initialize_wasm_instance(wasm_instance *instance);
 
 /**
  * Extract the data structure inside the `wasm_instance` resource.