--- a/extension/wasm.hh
+++ b/extension/wasm.hh
@@ -151,6 +151,23 @@ static void php_wasm_module_clean_up_persistent_resources();
 const char* wasm_instance_resource_name;
 int wasm_instance_resource_number;
 
+typedef struct {
+    // The internal opaque exported function pointer.
+    const wasmer_export_func_t *exported_function;
+
+    // The input signature of the exported function.
+    wasmer_value_tag *inputs;
+
+    // The input arity.
+    uint32_t input_arity;
+
+    // The output signature of the exported function.
+    wasmer_value_tag *outputs;
+
+    // The output ariry.
+    uint32_t output_arity;
+} wasm_exported_function;
+
 /**
  * Represents an `instance` with regular data.
  */
@@ -162,7 +179,7 @@ typedef struct {
     wasmer_exports_t *exports;
 
     // A map from exported function names to exported function pointers.
-    std::unordered_map<std::string, const wasmer_export_func_t *> *exported_functions;
+    std::unordered_map<std::string, wasm_exported_function *> *exported_functions;
 } wasm_instance;
 
 /**