The idea is to cache the exported functions of an instance into an
unordered map, so that we don't need to iterate over all the exports
each time to find an exported function to invoke.

`std::unordered_map` guarantees that reading an element is O(1).