@Hywan 2019-07-10
The `InvokeFunction` benchmark can help to measure the improvement here.

Before:

| subject                          | params | revs | its | mem_peak   | mean    | mode    | best    | rstdev |
|----------------------------------|--------|------|-----|------------|---------|---------|---------|--------|
| bench_invoke_sum                 | []     | 1000 | 10  | 1,421,000b | 8.169μs | 8.177μs | 7.752μs | 2.12%  |
| bench_invoke_i32_i64_f32_f64_f64 | []     | 1000 | 10  | 1,421,000b | 8.217μs | 8.143μs | 7.997μs | 2.92%  |

After:

| subject                          | params | revs | its | mem_peak   | mean    | mode    | best    | rstdev |
|----------------------------------|--------|------|-----|------------|---------|---------|---------|--------|
| bench_invoke_sum                 | []     | 1000 | 10  | 1,421,000b | 1.821μs | 1.799μs | 1.758μs | 3.77%  |
| bench_invoke_i32_i64_f32_f64_f64 | []     | 1000 | 10  | 1,421,000b | 1.983μs | 1.926μs | 1.875μs | 6.51%  |

It's 4.3 times faster to invoke a function (or 77% faster)! That's nice :-).